changing int to long.  Still won't compile.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@2992 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Cache/AccountC.cs
@@ -37,7 +37,7 @@ namespace OpenDentBusiness {
 		}
 
 		///<summary>Loops through listLong to find a description for the specified account.  0 returns an empty string.</summary>
-		public static string GetDescript(int accountNum){
+		public static string GetDescript(long accountNum){
 			if(accountNum==0) {
 				return "";
 			}
@@ -50,7 +50,7 @@ namespace OpenDentBusiness {
 		}
 
 		///<summary>Loops through listLong to find an account.  Will return null if accountNum is 0.</summary>
-		public static Account GetAccount(int accountNum) {
+		public static Account GetAccount(long accountNum) {
 			if(accountNum==0){
 				return null;
 			}
