+++ OpenDentBusiness/Cache/CovCatC.cs
@@ -35,7 +35,7 @@ namespace OpenDentBusiness {
 		}
 
 		///<summary></summary>
-		public static int GetOrderLong(int covCatNum) {
+		public static int GetOrderLong(long covCatNum) {
 			for(int i=0;i<Listt.Count;i++) {
 				if(covCatNum==Listt[i].CovCatNum) {
 					return i;
